minor bug corrections

+++ main/binary_mask_utils.py
@@ -59,7 +59,7 @@ def generate_coordinates_poly(poly):
     newPoly = [(x - minx, y - miny) for (x, y) in poly]
     X = maxx - minx + 1
     Y = maxy - miny + 1
-    grid = np.zeros((X, Y), dtype=np.int8)
+    grid = np.zeros((X, Y), dtype=np.int16)
     mahotas.polygon.fill_polygon(newPoly, grid)
     return [(x + minx, y + miny) for (x, y) in zip(*np.nonzero(grid))]
